moved the arrkeys function to its own library, because it may be somewhat usefull in other projects
