+++ game_main.c
@@ -5,6 +5,7 @@
 #include <time.h>
 #include <math.h>
 #include <string.h>
+#include "libs/arrkeys.h"
 #define MAX  23
 typedef struct axys{
     int x;
@@ -24,7 +25,7 @@ void posiSet(int x,int y,pos *a);
 void widewall(int x1, int x2, int y , pos *a,int *ind);
 void longwall(int y1, int y2, int x , pos *a, int *ind);
 void fightSmiley(chara *hero, int ind, int enN);
-char arrkeys();
+
 void enemy(char *ene, chara *bad, char *ename, int enN);
 void loadMap(pos *a, pos *b, pos *goal,int lv);
 void loadLv(int *nLv, int *enN);
@@ -405,31 +406,6 @@ void fightSmiley(chara *hero, int ind, int enN){
     }
 
 
-}
-
-char arrkeys(){
-    int ch = getch();
-    if (ch == 0 || ch == 224){
-        switch (_getch ()){
-            case 72:
-                return 'u';
-                break;
-            case 77:
-                return 'r';
-                break;
-            case 75:
-                return 'l';
-                break;
-            case 80:
-                return 'd';
-                break;
-            default:
-                return 'e';
-        }
-    }
-    if(ch == 13){
-        return 'q';
-    }
 }
 void loadMap(pos *a, pos *b, pos *goal, int lv){
     char ch;
