+++ libs/arrkeys.h
@@ -1,0 +1,6 @@
+#ifndef _ARRKEYS_LIB_
+#define _ARRKEYS_LIB_
+
+    char arrkeys();
+
+#endif // _ARRKEYS_LIB_
